@@ -9,18 +9,19 @@
     "axios": "^0.26.1",
     "bootstrap": "5.1.3",
     "bootstrap-social": "5.1.1",
-    "cross-fetch": "3.1.4",
+    "cross-fetch": "^3.1.5",
     "font-awesome": "4.7.0",
+    "framer-motion": "^7.3.5",
     "prop-types": "15.7.2",
-    "react": "17.0.2",
+    "react": "^18.2.0",
     "react-animation-components": "3.0.0",
-    "react-dom": "17.0.2",
+    "react-dom": "^18.2.0",
     "react-popper": "2.2.5",
     "react-redux": "7.2.6",
     "react-redux-form": "1.16.14",
     "react-router": "6.0.2",
-    "react-router-dom": "4.2.2",
-    "react-scripts": "4.0.3",
+    "react-router-dom": "^6.4.0",
+    "react-scripts": "^5.0.1",
     "react-transition-group": "4.4.2",
     "reactstrap": "8.10.1",
     "redux": "4.1.2",
