@@ -1,9 +1,9 @@
 import axios from 'axios'
 import {Component} from 'react'
 import {Link} from 'react-router-dom'
-import { Redirect } from 'react-router-dom'
+import { Navigate } from 'react-router-dom'
 import {connect} from 'react-redux'
-import {withRouter} from 'react-router-dom'
+import { motion } from "framer-motion"
 import './Register.css'
 
 const mapStateToProps = state => {
@@ -62,10 +62,18 @@ class Register extends Component{
     render(){
         return(
             <div>
-                <div>
+                <motion.div
+                    initial={{ opacity: 0 }}
+                    animate={{ opacity: 1 }}
+                    exit={{ opacity: 0 }}
+                >
                     <img src={signupBlob} alt="Yellow Blob" id="signupBlob" />
-                </div>
-                <div className='registerPage'>
+                </motion.div>
+                <motion.div className='registerPage'
+                    initial={{ left: "-300px", transition: { duration: .4 } }}
+                    animate={{ left: "50%", transition: { duration: .4, delay: .4 } }}
+                    exit={{ left: "-300px", transition: { duration: .4 }}}
+                >
                     <h1 className='please'>Create Account</h1>
                     <div className='user-box'>
                         <input
@@ -105,11 +113,11 @@ class Register extends Component{
                     </div>
                     <Link to="/login">Have an account?</Link>
                     <button type="submit" onClick={this.handleSubmit}>Create Account</button>
-                    {this.state.created ? <Redirect to='/login'/> : <></>}
-                </div>
+                    {this.state.created ? <Navigate to='/login'/> : <></>}
+                </motion.div>
             </div>
         )
     }
 }
 
-export default withRouter(connect(mapStateToProps)(Register));
+export default connect(mapStateToProps)(Register);
